Improve example safety
Since the example is not unlikely to be the basis for modified it is
worth to follow the guidelines closely. This removes some unnecessary or
slightly unsound blocks. Pointers can be default initialized with null
using the standard library.

And `mem::uninitialized` should under no circumstance be used as an
initialization, instead `mem::zeroed` should be used for types where all
bit representations are valid.